feat(GODT-2914): File Logging
Add file logging to the export tool. The logs are generated in the same
directory as the executable in the `logs` directory.

This patch also extends the constants with build time and git reversion
information.

Finally we also expose logging functions so that the C++ code can hook
into the Go log system.