feat(GODT-2919): Crash Recovery with Sentry Reports
When a go-routine panics, the recover action will log the stack trace to
the same log file and then call an external function to perform an
action on recovery.

If no recovery action is specified, the program will just exit. For the
CLI application, we print an error message to the user and then exit.