Support filtering backup by address
In Proton Mail you can have multiple addresses associated with an
account. To delete an address you must delete all messages associated
with that address. Previously the backup tool only supported exporting
emails from all addresses in the account. This PR adds the ability to
filter which address to export with the -a/--address command line
option. The default remains to export emails from all addreses.